fix unittest
This is quite an interesting commit. Unit test was failing if at least
one module was connected and security backend is Dummy. It happened
because dummy authorization token was not registered if app is imported
from other module.

To solve this problem, now application returned by
easy_phi.app.get_application(). This method also takes care of
registering dummy token